fix: return error instead of skipping inside of loop
Previously, we would continue if an error was returned when:
1. getting a transaction (or if there was no transaction)
2. getting an enqueued transaction (or if there was no enqueued transaction)
3. applying a transaction

Instead of silently skipping, we must always return an error.
There should be an explicit assumption on the DTL, that it will have
no gaps and that it can reliably provide us the data between [start, end].
If not, there should be an error since we want to favor safety over liveness.